# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prism",
    "version": "0.1.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/indexes": {
      "post": {
        "summary": "Create",
        "operationId": "create_v1_indexes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexes/{name}/stats": {
      "get": {
        "summary": "Stats",
        "operationId": "stats_v1_indexes__name__stats_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexes/{name}/documents:bulk": {
      "post": {
        "summary": "Bulk",
        "operationId": "bulk_v1_indexes__name__documents_bulk_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexes/{name}/documents/{external}": {
      "delete": {
        "summary": "Delete",
        "operationId": "delete_v1_indexes__name__documents__external__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "external",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexes/{name}/refresh": {
      "post": {
        "summary": "Refresh",
        "operationId": "refresh_v1_indexes__name__refresh_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexes/{name}/compact": {
      "post": {
        "summary": "Compact",
        "operationId": "compact_v1_indexes__name__compact_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexes/{name}/search": {
      "post": {
        "summary": "Search",
        "operationId": "search_v1_indexes__name__search_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BulkRequest": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Documents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "BulkRequest"
      },
      "CreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "semantic": {
            "type": "boolean",
            "title": "Semantic",
            "default": false
          },
          "codec": {
            "type": "string",
            "title": "Codec",
            "default": "compressed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateRequest"
      },
      "Document": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "kind": {
            "type": "string",
            "enum": [
              "product",
              "service",
              "document"
            ],
            "title": "Kind",
            "default": "document"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 64,
            "title": "Tags"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "Document"
      },
      "Filter": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "op": {
            "type": "string",
            "enum": [
              "eq",
              "in",
              "gt",
              "gte",
              "lt",
              "lte"
            ],
            "title": "Op",
            "default": "eq"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "value"
        ],
        "title": "Filter"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RefreshRequest": {
        "properties": {
          "through_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Through Revision"
          }
        },
        "type": "object",
        "title": "RefreshRequest"
      },
      "SearchRequest": {
        "properties": {
          "q": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Q"
          },
          "query_mode": {
            "type": "string",
            "enum": [
              "plain",
              "dsl"
            ],
            "title": "Query Mode",
            "default": "plain"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "lexical",
                  "semantic",
                  "hybrid"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "filters": {
            "items": {
              "$ref": "#/components/schemas/Filter"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Filters"
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "candidate_k": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Candidate K",
            "default": 100
          },
          "typo_tolerance": {
            "type": "boolean",
            "title": "Typo Tolerance",
            "default": true
          },
          "synonyms": {
            "type": "boolean",
            "title": "Synonyms",
            "default": true
          },
          "timeout_ms": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 10.0,
            "title": "Timeout Ms",
            "default": 500
          },
          "allow_partial": {
            "type": "boolean",
            "title": "Allow Partial",
            "default": true
          },
          "allow_lexical_fallback": {
            "type": "boolean",
            "title": "Allow Lexical Fallback",
            "default": false
          },
          "explain": {
            "type": "boolean",
            "title": "Explain",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "q"
        ],
        "title": "SearchRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}